// src/types/category.d.ts
export type CategoryLabel =
  | "전체"
  | "디지털기기"
  | "생활가전"
  | "가구/인테리어"
  | "생활/주방"
  | "유아동"
  | "유아도서"
  | "여성의류"
  | "여성잡화"
  | "남성패션/잡화"
  | "뷰티/미용"
  | "스포츠/레저"
  | "취미/게임/음반";